private methods where possible. move code to ranks. fixing rank update bug. adding sql scripts
